+++ packages/components/inline-edit/index.ts
@@ -1,3 +1,2 @@
-export * from './focus-monitor';
 export * from './inline-edit';
 export * from './module';
