--- packages/components/inline-edit/mask.scss
@@ -1,37 +1,0 @@
-@use 'sass:map';
-
-@mixin mask($config) {
-    & {
-        position: relative;
-    }
-
-    .kbq-mask {
-        display: flex;
-        position: absolute;
-
-        @if (map.get($config, 'top')) {
-            top: map.get($config, 'top');
-        }
-
-        @if (map.get($config, 'right')) {
-            right: map.get($config, 'right');
-        }
-
-        @if (map.get($config, 'bottom')) {
-            bottom: map.get($config, 'bottom');
-        }
-
-        @if (map.get($config, 'left')) {
-            left: map.get($config, 'left');
-        }
-
-        .kbq-mask__fade {
-            background: var(--mask-background, transparent);
-            mask-image: linear-gradient(90deg, transparent, var(--kbq-black-default));
-        }
-
-        .kbq-mask__container {
-            background: var(--mask-background, transparent);
-        }
-    }
-}
